requirements.txt
Flask_Login==0.5.0
Flask==2.0.1
flask_sqlalchemy==2.5.1
python-dotenv==0.20.0
python_bcrypt==0.3.2
